fix(day-8): correctness bugs, task registry, download endpoint, structured logging
import missing timezone (blank created_at crashed the whole job)
money as Decimal end to end, not float
TASK_REGISTRY so job_type dispatches the right task; 400 on unknown
add GET /jobs/{id}/download; reports were unreachable on a volume
count CSV rows with csv.reader (quoted newlines miscounted)
drop the fake time.sleep throttle and Base.metadata.create_all
structlog with request_id/job_id/task_id correlation; SQL_ECHO off DEBUG
lazy sync engine so tests can rebind DATABASE_URL

+++ backend/app/models.py
@@ -5,8 +5,9 @@ SQLAlchemy ORM models for the Async Job Processing Platform.
 import enum
 import uuid
 from datetime import datetime, timezone
+from decimal import Decimal
 
-from sqlalchemy import DateTime, Enum, Index, Integer, Numeric, String, Text, text
+from sqlalchemy import DateTime, Enum, Index, Integer, Numeric, String, text
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -91,7 +92,8 @@ class Transaction(Base):
         nullable=False,
         index=True,
     )
-    amount: Mapped[float] = mapped_column(
+    # Money: Numeric(12, 2) maps to Decimal, never float. See CLAUDE.md.
+    amount: Mapped[Decimal] = mapped_column(
         Numeric(12, 2),
         nullable=False,
     )
